--- GraviTrax Connect Python Library/MANIFEST.in.txt
@@ -1,4 +1,0 @@
-include LICENSE.txt
-include README.md
-include CHANGES.md
-include docs/*
