+++ GraviTrax-Connect-Python-Library/CHANGES.md
@@ -1,0 +1,4 @@
+# Gravitraxconnect Changelog
+
+## 2023-06-01  Version 0.3.2
+- Initial Release
