+++ GraviTrax-Connect-Python-Library/MANIFEST.in.txt
@@ -1,0 +1,4 @@
+include LICENSE.txt
+include README.md
+include CHANGES.md
+include docs/*
